+++ src/castillo_chan_zhou_decrypt_binary/__init__.py
@@ -1,5 +1,5 @@
-"""Top-level package for CS6903-Project-One."""
+"""Top-level package for castillo-chan-zhou-decrypt-binary."""
 
-__author__ = """Joel Castillo"""
-__email__ = "joel.castillo@nyu.edu"
-__version__ = "0.1.0"
+__authors__ = """Joel Castillo, Ho Yin Kenneth Chan, Gary Zhou"""
+__author_emails__ = """joel.castillo@nyu.edu, hyc449@nyu.edu, gz438@nyu.edu"""
+__version__ = "1.0.0"