+++ CHANGELOG.md
@@ -7,9 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.1.0] - 2021-03-06
+## [1.0.0] - 2021-03-19
 ### Added
-- First release on PyPI.
+- CLI for Test One and Test Two of CS6903 Project.
+- Included report
+- Updated documentation
 
-[Unreleased]: https://github.com/joelbcastillo/CS6903-Project-One/compare/v0.1.0...HEAD
-[0.1.0]: https://github.com/joelbcastillo/CS6903-Project-One/compare/releases/tag/v0.1.0
+## [0.1.0] - 2021-03-19
+
+### Added
+- Test release pipeline
+
+[Unreleased]: https://github.com/joelbcastillo/castillo-chan-zhou-decrypt-binary/compare/v1.0.0...HEAD
+[1.0.0]: https://github.com/joelbcastillo/castillo-chan-zhou-decrypt-binary/compare/releases/tag/v1.0.0
+[0.1.0]: https://github.com/joelbcastillo/castillo-chan-zhou-decrypt-binary/compare/releases/tag/v0.1.0